# api-cms/docs/swagger.yaml
definitions:
...
  admin.Label:
    properties:
      name:
        type: string
      slug:
        type: string
    type: object
  admin.LinkNext:
    properties:
      box_element:
        type: string
      category_element:
        type: string
      check_category:
        type: string
      check_desc:
        type: string
      check_time:
        type: string
      content_element:
        type: string
      description_element:
        type: string
      link_element:
        type: string
      time_element:
        type: string
      title_element:
        type: string
      url_start:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/common/labels:
    get:
      consumes:
      - application/json
      description: Xem danh sách Label
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem danh sách Label
      tags:
      - Label
    post:
      consumes:
      - application/json
      description: Tạo mới Label
      parameters:
      - description: name
        in: body
        name: LabelForm
        required: true
        schema:
          $ref: '#/definitions/admin.Label'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Tạo mới Label
      tags:
      - Label
  /api/v1/common/labels/{label_id}:
    delete:
      consumes:
      - application/json
      description: Xóa Label
      parameters:
      - description: label_id
        in: path
        name: label_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xóa Label
      tags:
      - Label
    get:
      consumes:
      - application/json
      description: Xem chi tiết Label
      parameters:
      - description: label_id
        in: path
        name: label_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem chi tiết Label
      tags:
      - Label
    put:
      consumes:
      - application/json
      description: Sửa Label
      parameters:
      - description: label_id
        in: path
        name: label_id
        required: true
        type: string
      - description: Thông tin cập nhật của Label
        in: body
        name: LabelForm
        required: true
        schema:
          $ref: '#/definitions/admin.Label'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update Label
      tags:
      - Label
  /api/v1/common/linkNexts:
    get:
      consumes:
      - application/json
      description: Xem danh sách LinkNext
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem danh sách LinkNext
      tags:
      - LinkNext
    post:
      consumes:
      - application/json
      description: Tạo mới LinkNext
      parameters:
      - description: Name
        in: body
        name: LinkNextForm
        required: true
        schema:
          $ref: '#/definitions/admin.LinkNext'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Tạo mới LinkNext
      tags:
      - LinkNext
  /api/v1/common/linkNexts/{linkNext_id}:
    delete:
      consumes:
      - application/json
      description: Xóa LinkNext
      parameters:
      - description: linkNext_id
        in: path
        name: linkNext_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xóa LinkNext
      tags:
      - LinkNext
    get:
      consumes:
      - application/json
      description: Xem chi tiết LinkNext
      parameters:
      - description: linkNext_id
        in: path
        name: linkNext_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem chi tiết LinkNext
      tags:
      - LinkNext
    put:
      consumes:
      - application/json
      description: Cập nhật LinkNext
      parameters:
      - description: linkNext_id của LinkNext cần cập nhật
        in: path
        name: linkNext_id
        required: true
        type: string
      - description: Thông tin cập nhật của LinkNext
        in: body
        name: LinkNextForm
        required: true
        schema:
          $ref: '#/definitions/admin.LinkNext'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Cập nhật LinkNext
      tags:
      - LinkNext
...
swagger: "2.0"